backend/requirements.txt
flask==2.3.3
groq==0.4.2
python-dotenv==1.0.0
flask-cors==4.0.0
